Revert to state before refactoring the forcing
