+++ ewatercycle/models/wflow.py
@@ -12,9 +12,48 @@ from grpc4bmi.bmi_client_docker import BmiClientDocker
 from grpc4bmi.bmi_client_singularity import BmiClientSingularity
 
 from ewatercycle import CFG
-from ewatercycle.forcing.forcing_data import ForcingData
 from ewatercycle.models.abstract import AbstractModel
 from ewatercycle.parametersetdb.config import CaseConfigParser
+
+class PathParser:
+    """Descriptor that converts input to pathlib.Path objects."""
+    def __set_name__(self, owner: object, name: str):
+        self.name = name
+
+    def __set__(self, instance: object, value: Union[str, PathLike]):
+        instance.__dict__[self.name] = Path(value).expanduser().resolve()
+
+
+@dataclass
+class WflowForcing:
+    """Forcing data for the Wflow model class.
+
+    Default variable names follow CMOR standards.
+
+    Example:
+
+        To run the example case `wflow_rhine_sbm_nc <https://github.com/openstreams/wflow/tree/master/examples/wflow_rhine_sbm_nc>`_ one would use:
+
+    .. code-block::
+
+        forcing = WflowForcing(
+            netcdfinput='inmaps.nc',
+            Precipitation = "/P",
+            EvapoTranspiration = "/PET",
+            Temperature = "/TEMP",
+        )
+    """
+    netcdfinput: PathParser = PathParser()
+    """Input file path."""
+    Precipitation: str = "/pr"
+    """Variable name of Precipitation data in input file."""
+    EvapoTranspiration: str = "/pet"
+    """Variable name of EvapoTranspiration data in input file."""
+    Temperature: str = "/tas"
+    """Variable name of Temperature data in input file."""
+    Inflow: Optional[str] = None
+    """Variable name of Inflow data in input file."""
+
 
 @dataclass
 class WflowParameterSet:
@@ -50,7 +89,7 @@ class Wflow(AbstractModel):
     def __init__(self,
                  version: str,
                  parameter_set: WflowParameterSet,
-                 forcing: ForcingData):
+                 forcing: Optional[WflowForcing] = None):
 
         super().__init__()
         self.version = version
@@ -85,13 +124,16 @@ class Wflow(AbstractModel):
         self.config = cfg
 
     def _parse_forcing(self):
+        if self.forcing is None:
+            return
+
         cfg = self.config
         forcing = self.forcing
-        cfg.set("framework", "netcdfinput", forcing.configuration.netcdfinput)
-        cfg.set("inputmapstacks", "Precipitation", forcing.configuration.Precipitation)
-        cfg.set("inputmapstacks", "Temperature", forcing.configuration.Temperature)
+        cfg.set("framework", "netcdfinput", forcing.netcdfinput.name)
+        cfg.set("inputmapstacks", "Precipitation", forcing.Precipitation)
         cfg.set("inputmapstacks", "EvapoTranspiration",
-                forcing.configuration.EvapoTranspiration)
+                forcing.EvapoTranspiration)
+        cfg.set("inputmapstacks", "Temperature", forcing.Temperature)
 
     def setup(self, **kwargs) -> Tuple[PathLike, PathLike]:  # type: ignore
         """Start the model inside a container and return a valid config file.
